// src/data/works-images.generated.ts
export type WorkImage = {
  slug: string;
  src: string;
  width: number;
  height: number;
  blurDataURL: string;
};

export const workImages: Record<string, WorkImage> = {
  "red-knott": {
    "slug": "red-knott",
    "src": "/works/red-knott.avif",
    "width": 1125,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRn4AAABXRUJQVlA4IHIAAACwAwCdASoLABQAPu1iqU2ppaQiMAgBMB2JQAALd424/SI3yyagWAD+77wk6nU6f1vktRLTEVS+J43r9y3DqkvlmdvXERABl6elRMnwHzLIYPPwCelqOyHjOcn8Tqv2fOVk8XsmMma181UfVzq8Zfx0IAA="
  },
  "heritage-board": {
    "slug": "heritage-board",
    "src": "/works/heritage-board.avif",
    "width": 823,
    "height": 1097,
    "blurDataURL": "data:image/webp;base64,UklGRmAAAABXRUJQVlA4IFQAAACwAwCdASoPABQAPu1iqU2ppaOiMAgBMB2JQAALg3FGy6oZy9DhwAD+7/t0BPg9TAO2C/SF9j/skJXp5KfSxHb5sF2y+dzJZZgxJcWZH1Jykj0AAAA="
  },
  "festive-favourites": {
    "slug": "festive-favourites",
    "src": "/works/festive-favourites.avif",
    "width": 1125,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRpoAAABXRUJQVlA4II4AAAAwBACdASoLABQAPu1iqU2ppaOiMAgBMB2JQBOmUABV1OofHmN1eQdWsRAA/u1hd5K9Lvv9D5TbfKjS7fRpiiNSh1LFQm/stuW3F3fn4Ks3wFAmMFHx8P0jn1WyA32CzjmTrooWCOXtM3xCTMaoZQgqi5KWZ1FMaQ7Y5lIJwJBVx9m8+xUP6E3WVEHPAAAA"
  },
  "birthday-sale": {
    "slug": "birthday-sale",
    "src": "/works/birthday-sale.avif",
    "width": 2000,
    "height": 1000,
    "blurDataURL": "data:image/webp;base64,UklGRoAAAABXRUJQVlA4IHQAAACQAwCdASoUAAoAPu1iqU2ppaOiMAgBMB2JQBOgAlbbKQYpd9ugAMl2ewVZjt68k+T46oHH6PxyYcNYHhZ9hvHnYqLk5wB3RrRZDd/hMnnFvsLZYHW4h8/9Xmywwj4BnxUpni7ig4YDjOp48oGbpXhMHG2gAA=="
  },
  "hoodie-mockup": {
    "slug": "hoodie-mockup",
    "src": "/works/hoodie-mockup.avif",
    "width": 2000,
    "height": 1333,
    "blurDataURL": "data:image/webp;base64,UklGRnAAAABXRUJQVlA4IGQAAADQAwCdASoUAA0APu1iqk2ppaQiMAgBMB2JQBYdgwv5HT+Fh1v/MPgA/nGaTwbR3v30tCpbJPJKbf1CIiMFPIagnZNNylY7etpphyPXbIsm0hqxa/stpeu7YPqRGV/kx7c9rrgA"
  },
  "snakes-and-ladders": {
    "slug": "snakes-and-ladders",
    "src": "/works/snakes-and-ladders.avif",
    "width": 1500,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRpoAAABXRUJQVlA4II4AAABwBACdASoPABQAPu1iqU2ppaOiMAgBMB2JagCxG1AViUuruIR8VQxQy/j7AAD+0O9db9/OKAny/RB0TNaUgbTfbzpGaUEBqrojtP/x0NE8Oz/F4c9bFyhGt6B+KgDCJi3WXFF2IX/lBCWOvkqWtLkciTcOatetvXj5c+eByNimNiZOvrOaP/4KldhjZ8wA"
  },
  "peace": {
    "slug": "peace",
    "src": "/works/peace.avif",
    "width": 1920,
    "height": 1080,
    "blurDataURL": "data:image/webp;base64,UklGRlwAAABXRUJQVlA4IFAAAABwAwCdASoUAAsAPu1iqU2ppaOiMAgBMB2JYgC7AB4XW1b5CAAA/ujxJur24tRjJ1owgICYICXvMHmoWYQni7BGJ3/Hh9QiDP3sleAACMAAAA=="
  },
  "infographic": {
    "slug": "infographic",
    "src": "/works/infographic.avif",
    "width": 768,
    "height": 1024,
    "blurDataURL": "data:image/webp;base64,UklGRmIAAABXRUJQVlA4IFYAAACwAwCdASoPABQAPu1kqk4ppaQiMAgBMB2JQBadBD4ZutnOViOiwAD+7/t0BPg9S4tkEE8CsGSKKq6kdNeW0dqcwgR+dRcjwDUj2EGdmdJqyoehFx9AAA=="
  },
  "campaign-ten": {
    "slug": "campaign-ten",
    "src": "/works/campaign-ten.avif",
    "width": 2000,
    "height": 1000,
    "blurDataURL": "data:image/webp;base64,UklGRoYAAABXRUJQVlA4IHoAAADwAwCdASoUAAoAPu1iqU2ppaOiMAgBMB2JbACdMoRwABM4zWUTV+IkAPyLfjTumKruNJ3TYRaWIAH86AZYxejDEE7xWomolpWtTWAsfiNcWZ6kYpq7dFlK6K9G1afonn9uWaTzISQ71tDIhhfP17vEZsOxzT7DuEgAAA=="
  },
  "editorial-spread": {
    "slug": "editorial-spread",
    "src": "/works/editorial-spread.avif",
    "width": 2000,
    "height": 1000,
    "blurDataURL": "data:image/webp;base64,UklGRmgAAABXRUJQVlA4IFwAAADQAwCdASoUAAoAPu1iqU2ppaOiMAgBMB2JYwC06Buzs69GOGJKiAAA/q7Y5Hv3XL7kxipowPkLsaliFEswcsBJcIbCBIjyDsqutB20go3SoEt85zu1c9BG/dAAAA=="
  },
  "editorial-spread-two": {
    "slug": "editorial-spread-two",
    "src": "/works/editorial-spread-two.avif",
    "width": 2000,
    "height": 1000,
    "blurDataURL": "data:image/webp;base64,UklGRmgAAABXRUJQVlA4IFwAAADQAwCdASoUAAoAPu1iqU2ppaOiMAgBMB2JYwC06Buzs69GOGJKiAAA/q7Y5Hv3XL7kxipowPkLsaliFEswcsBJcIbCBIjyDsqutB20go3SoEt85zu1c9BG/dAAAA=="
  },
  "identity-study": {
    "slug": "identity-study",
    "src": "/works/identity-study.avif",
    "width": 2000,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRrgAAABXRUJQVlA4IKwAAADwBACdASoUABQAPu1ur1IppiQiqAgBMB2JQBYj6rgcne5fui+3avXXlEyu3O1M0wAA/pTL4N39f480l2GbfGhkEzg0JMcqCPzGDNvgfMRpz24XxT89Mr2lphtP/YuQUArQEaWOwyuq72s+juEBeY03716m87iioP8I09Dp454ioWmBB28EgQWqeoj19DsyDJ5A9aW08ku9AZ1ebV3RQ9mbdUZ5+fEG+MsEAAAA"
  },
  "identity-study-two": {
    "slug": "identity-study-two",
    "src": "/works/identity-study-two.avif",
    "width": 1125,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRoYAAABXRUJQVlA4IHoAAAAQBACdASoLABQAPu1iqU2ppaOiMAgBMB2JYwCdMoAlqRMsAg8SEf51gAD5YR/JIl6BvHzG7gA9uguA4oUPFZwFnEofoLfnEjqlmhx2dVrudgl4vDpeOyf1ftSsrQ/CDcpsE+aob9igDh3mf2wOcYKCN4vPXk6GHAdAAA=="
  }
};